Develop

See merge request s08-blockchain-contract-sub2/S08P22A209!274

+++ backend/src/main/java/com/donjo/backend/api/dto/member/response/SupportItem.java
@@ -27,7 +27,7 @@ public class SupportItem {
   // 보낸 지갑 주소
   private ToMemberItem toMember;
   // 보낸 이더
-  private Long amountEth;
+  private Double amountEth;
   // 보낸 시간
   private LocalDateTime sendTimeStamp;
   // 도착 시간
